+rearrange module with math
