+++ src/math/mod.rs
@@ -1,0 +1,16 @@
+pub mod mean_colmogorov;
+pub(crate) mod mean_winsorized;
+pub mod median;
+pub mod mean_arithmetic_geometric_modified;
+pub mod mean_arithmetic;
+pub mod mean_arithmetic_geometric;
+pub mod mean_truncated;
+pub mod mode;
+pub mod mean_geometric;
+pub mod mean_generalized;
+pub mod average_absolute_deviation;
+pub mod standard_variation_coefficient;
+pub mod standard_deviation;
+pub mod linear_variation_coefficient;
+pub mod variance;
+
